Add native java screen recorder (1/2)
This is based on the screenrecorder binary found in frameworks/av.
It has been converted to a native library with the necessary jni
hooks for interfacing withit via a ScreenRecorder java class.

Volume up + power chord is used to start and stop recording by
binding to the service found in the ScreenRecorder app:
https://github.com/ChameleonOS/android_packages_apps_ScreenRecorder

Change-Id: I5b7af9b4b4f24a8c54ac0f9dcb36ace16672e2cf
Signed-off-by: Almite <almite@hotmail.com>